third_party/WebKit/LayoutTests/media/track/track-id.html

<!DOCTYPE html>
<title>Tests that the TextTrack "id" attribute is set appropriately.</title>
<video>
    <track id="LoremIpsum" src="captions-webvtt/captions-fast.vtt" default>
</video>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
    var video = document.querySelector("video");

    var track = document.querySelector("track");
    
    var textTrack = track.track;

    // Test default attribute value.
    assert_equals(textTrack.id, "LoremIpsum");
    assert_equals(video.textTracks[0].id, "LoremIpsum");

    // Make sure we can look up tracks by id.
    assert_equals(video.textTracks.getTrackById("LoremIpsum"), textTrack);

    // Test that it's readonly.
    textTrack.id = "newvalue";
    assert_equals(textTrack.id, "LoremIpsum");
});
</script>
